Update patch set 8
Patch Set 8:

> Consider the following case, we have three branches b, c and d, and
 > we have the maxBranches set to 1, so the branch b will be
 > monitored, now we add a new branch a, and reload the plugin, which
 > branch will be monitored now?
I see what you are saying and that is a problem.  I am just thinking out loud in terms of the possible unboundedness of this new functionality.  Another possibility is adding regex support to the branch spec?  Although one can still get a large number of branches from that.

With a configurable max, we can potentially log an error when the max is reached.  Don't know if that's the best way to handle it though.

Patch-set: 8
Reviewer: Gerrit User 1021862 <1021862@173816e5-2b9a-37c3-8a2e-48639d4f1153>